# perf(marshal): Extend new assert style to passStyleOf `check` calls