--- a/packages/marshal/src/helpers/copyRecord.js
+++ b/packages/marshal/src/helpers/copyRecord.js
@@ -24,9 +24,10 @@ export const CopyRecordHelper = harden({
   styleName: 'copyRecord',
 
   canBeValid: (candidate, check) => {
+    const reject = details => check(false, details);
     const proto = getPrototypeOf(candidate);
     if (proto !== objectPrototype && proto !== null) {
-      return check(false, X`Unexpected prototype for: ${candidate}`);
+      return reject(X`Unexpected prototype for: ${candidate}`);
     }
     const descs = getOwnPropertyDescriptors(candidate);
     const descKeys = ownKeys(descs);
@@ -34,15 +35,13 @@ export const CopyRecordHelper = harden({
     for (const descKey of descKeys) {
       if (typeof descKey !== 'string') {
         // Pass by copy
-        return check(
-          false,
+        return reject(
           X`Records can only have string-named own properties: ${candidate}`,
         );
       }
       const desc = descs[descKey];
       if (canBeMethod(desc.value)) {
-        return check(
-          false,
+        return reject(
           X`Records cannot contain non-far functions because they may be methods of an implicit Remotable: ${candidate}`,
         );
       }