Ref dc24f2f2c3cac7ce239a64c503493c41a2334315

<details><summary>ad hoc analysis</summary>

```sh
$ node --input-type=module -e '
  import Benchmark from "benchmark";
  import "../init/index.js";
  import { passStyleOf } from "./src/passStyleOf.js";
  import { makeTagged } from "./src/makeTagged.js";
  const inputs = [
    undefined,
    "foo",
    true,
    33,
    33n,
    Symbol.for("foo"),
    Symbol.iterator,
    null,
    Promise.resolve(null),
    [3, 4],
    { foo: 3 },
    { then: "non-function then ok" },
    makeTagged("unknown", undefined),
    Error("ok"),
  ].map(v => harden(v));
  const badInputs = [
    Symbol("unique"),
    {},
    harden(Object.setPrototypeOf(Promise.resolve(), { __proto__: Promise.prototype })),
    harden(Object.assign(Promise.resolve(), { extra: "unexpected own property" })),
    harden(Object.defineProperties(Promise.resolve(), { then: { value: () => "bad then" } })),
    harden({ then: () => "thenable" }),
  ];
  let result;
  (new Benchmark.Suite())
    .add("valid inputs", () => {
      for (const v of inputs) result = passStyleOf(v);
    })
    .add("invalid inputs", () => {
      let caught;
      for (const v of badInputs) {
        caught = undefined;
        try {
          result = passStyleOf(v);
        } catch (err) {
          caught = err;
        } finally {
          if (!caught) throw new Error(`unexpected acceptance: ${v}`);
        }
      }
    })
    .add("big input", () => {
      result = passStyleOf(harden(Array(1000).fill().map(() => [...inputs])));
    })
    .on("cycle", evt => console.log(String(evt.target)))
    .on("complete", () => result)
    .run({ async: true })
'
```

</details>

before:
```
valid inputs x 831,564 ops/sec ±0.71% (91 runs sampled)
invalid inputs x 3,645 ops/sec ±0.86% (88 runs sampled)
big input x 10.31 ops/sec ±3.22% (32 runs sampled)
```
after:
```
valid inputs x 828,877 ops/sec ±0.60% (91 runs sampled)
invalid inputs x 3,780 ops/sec ±0.76% (85 runs sampled)
big input x 66.53 ops/sec ±1.74% (67 runs sampled)
```

(and we could probably get even more improvement by replacing `check` parameters with `reject` to avoid creating the latter functions inline)