@gibson042 2022-10-07
> Should I review this assuming that you intend it to be a pure refactoring --- with no observable differences? FWIW, I have not spotted any yet.

Correct. :+1: